feat(memory): add long-term user memory with E5 embedding
- Add UserMemoryService: Qdrant user-memories (384-dim, Cosine)
- Add MemoryExtractionService: async fact extraction via flash-lite
- Add MemoryRetentionAdvisor: StreamAdvisor post-stream extraction
- Add MemoryTools: recall_memories agentic tool for user context
- Add AugmentedToolCallbackProvider: reasoning logger for tools
- Add LocalEmbeddingConfig: multilingual-e5-small ONNX wrappers
- Migrate embedding: MiniLM-L6-v2 to multilingual-e5-small
- Upgrade Spring AI 1.1.0 to 1.1.4, pin ONNX Runtime 1.19.2
- Replace RequiredArgsConstructor with explicit constructors
- Update docker-compose.prod: E5 volume, RAM 768 to 1024MB
- Update README: Memory docs, E5 download, compact tuning

+++ src/main/java/dev/thanh/spring_ai/service/SessionActivityService.java
@@ -16,7 +16,12 @@ import java.time.Duration;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
-import java.util.*;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 /**
@@ -59,7 +64,7 @@ public class SessionActivityService {
      * round-trip.
      */
     public void touchSession(String userId, String sessionId) {
-        double score = (double) System.currentTimeMillis();
+        double score = System.currentTimeMillis();
         String userKey = buildUserKey(userId);
 
         safeRedis.tryExecute(
@@ -177,7 +182,7 @@ public class SessionActivityService {
         return safeRedis.executeOrReject(
                 () -> {
                     Set<ZSetOperations.TypedTuple<Object>> tuples = redisTemplate.opsForZSet()
-                            .reverseRangeWithScores(userKey, 0, limit - 1);
+                            .reverseRangeWithScores(userKey, 0, (long) limit - 1);
 
                     if (tuples == null || tuples.isEmpty()) {
                         return Collections.<String>emptyList();
